% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATEDdetectPulses.R
\name{DEPRECATDdetectPulses}
\alias{DEPRECATDdetectPulses}
\title{Test for pulses using using Foote Rates, using ultrametric trees simulated using phytools::pbtree()
not currently recommended}
\usage{
DEPRECATDdetectPulses(myTree, criterion = 1.5, desiredBinNumber = 10,
  plotTree = FALSE, plotRates = FALSE, returnRatePlot = FALSE)
}
\arguments{
\item{myTree}{simulated phylogenetic tree}

\item{criterion}{How extreme an observed turnover rate must be to be considered a turnover pulse (relative to the interquartile distance for the distribution of all bins).}

\item{desiredBinNumber}{The number of bins for computing turnover rates.}

\item{plotTree}{Whether or not to plot the simulated tree. Default is FALSE.}

\item{plotRates}{Whether or not to plot a histogram of the calculated rates.  Default is TRUE.}

\item{returnRatePlot}{Whether or not to add the ggplot of the rates to the return value.}
}
\description{
Test for pulses using using Foote Rates, using ultrametric trees simulated using phytools::pbtree()
not currently recommended
}

