% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/DetectFADPulses.R
\name{DetectFADPulses}
\alias{DetectFADPulses}
\title{Bin FADs from a tree, and test for pulses using chi-square}
\usage{
DetectFADPulses(nBins, myTree, showPlot = FALSE)
}
\arguments{
\item{nBins}{the desired number of bins for combining FADs}

\item{myTree}{an object of class "phylo"}

\item{showPlot}{Whether or not to show plots. Default is FALSE}
}
\description{
Bin FADs from a tree, and test for pulses using chi-square
}

