% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPulses.R
\name{detectPulses}
\alias{detectPulses}
\title{test for pulses using paleotree functions}
\usage{
detectPulses(fossilRecords, binLength = 0.25, criterion = 1.5,
  plot = FALSE)
}
\arguments{
\item{fossilRecords}{an object of class fossilRecordPlusParams, produced by NullTurnover::simRecords() function}

\item{binLength}{Length of time corresponding to bins, in simulation time units. Default is 0.25}

\item{criterion}{Criterion for determining if a bin is a turnover pulse or not. Based on upper 3rd quartile of turnover rates for each bin.  Default 1.5 means that a pulse is defined as 1.5 times higher than the 3rd quartile, including zeros.}

\item{plot}{Whether or not to plot the ranges using NullTurnover::plotRanges(). Defaults to FALSE}
}
\description{
test for pulses using paleotree functions
}

