% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectPulsesChiSq.R
\name{detectPulsesChiSq}
\alias{detectPulsesChiSq}
\title{Test for pulses using using chisquare}
\usage{
detectPulsesChiSq(treeDepth = 7, criterion = 1.5, nTaxa = 100,
  deathRate = 0, desiredBinNumber = 30, plotTree = FALSE,
  plotRates = FALSE, excludeZeroes = FALSE)
}
\arguments{
\item{treeDepth}{The temporal depth (in millions of years) represented by the tree.}

\item{nTaxa}{The desired number of taxa in the simulated phylogenetic tree. Passed to phytools::pbtree.}

\item{deathRate}{The rate of extinction.  Passed to phytools::pbtree.}

\item{desiredBinNumber}{The number of bins that treeDepth will be broken into to compute turnover rates.}

\item{plotTree}{Whether or not to plot the simulated tree. Default is FALSE.}

\item{plotRates}{Whether or not to plot a histogram of the calculated rates.  Default is TRUE.}

\item{excludeZeroes}{Whether or not to exclude time bins with a rate of zero. Default is FALSE.}
}
\description{
Test for pulses using using chisquare
}

