% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/LoessDetectPulses.R
\name{LoessDetectPulses}
\alias{LoessDetectPulses}
\title{Test for pulses using residuals from loess}
\usage{
LoessDetectPulses(nBins, myTree = NULL, FADs = NULL, showPlots = FALSE,
  span = 0.2, showTree = FALSE)
}
\arguments{
\item{nBins}{the desired number of bins for combining FADs}

\item{myTree}{an object of class "phylo". You must include EITHER a value for myTree OR a value for FADs, but not both.}

\item{FADs}{Vector of taxon first appearances. You must include EITHER a value for myTree OR a value for FADs, but not both.}

\item{showTree}{Whether or not to plot the tree. Default is FALSE}

\item{showBinPlot}{Whether or not to show the plots of binned residuals. Default is FALSE}

\item{showExpected}{whether or not to show the plot of the expected versus actual.  Default is FALSE}
}
\description{
Test for pulses using residuals from loess
}

