\name{NullTurnover-package}
\alias{NullTurnover-package}
\alias{NullTurnover}
\docType{package}
\title{
Null Models For Turnover Patterns
}
\description{
This package uses simulations to create a null distribution for faunal turnover parameters
based on user-defined parameters.
}
\details{
\tabular{ll}{
Package: \tab NullTurnover\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-06-25\cr
License: \tab MIT + file LICENSE\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
W. Andrew Barr

Maintainer: W. Andrew Barr
}
\references{
Vrba. 1995. The Fossil Record of African Antelopes (Mammalia, Bovidae) in Relation to Human Evolution and Paleoclimate. In: Vrba ES, Denton GH, Partridge TC, Burckle LH, editors. Paleoclimate and Evolution with Emphasis on Human Origins. New Haven: Yale University Press. p 383–424.
}

\examples{
}
