% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRecords.R
\name{simRecords}
\alias{simRecords}
\title{simulate fossil records.  This function is a wrapper for paleotree::simFossilRecord that does error catching}
\usage{
simRecords(origRateSim, extRateSim, nruns, totalTime, sampRateSim, startTaxa,
  maxAttempts, nSamp = c(0, 500), nExtant = c(10, 500), nTotalTaxa = c(10,
  500), plot = TRUE, print.runs = FALSE)
}
\value{
a list of class fossilRecordPlusParams which includes the simulated record (or an error), plus a list of parameters used in the simulation.
}
\description{
simulate fossil records.  This function is a wrapper for paleotree::simFossilRecord that does error catching
}

