% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/TestFADs.R
\name{TestFADs}
\alias{TestFADs}
\title{Bin FADs from a tree, and test for pulses using chi-square}
\usage{
TestFADs(nBins, myTree)
}
\arguments{
\item{nBins}{the desired number of bins for combining FADs}

\item{myTree}{an object of class "phylo"}
}
\description{
Bin FADs from a tree, and test for pulses using chi-square
}

