% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/VrbaDetectPulses.R
\name{VrbaDetectPulses}
\alias{VrbaDetectPulses}
\title{Bin FADs from a tree, and test for pulses using chi-square, following Vrba (1995)}
\usage{
VrbaDetectPulses(nBins, myTree = NULL, FADs = NULL, showBinPlot = FALSE,
  nls_mod = NULL, showExpected = FALSE, showTree = FALSE)
}
\arguments{
\item{nBins}{the desired number of bins for combining FADs}

\item{myTree}{an object of class "phylo". You must include EITHER a value for myTree OR a value for FADs, but not both.}

\item{FADs}{Vector of taxon first appearances. You must include EITHER a value for myTree OR a value for FADs, but not both.}

\item{showBinPlot}{Whether or not to show the plots of binned residuals. Default is FALSE}

\item{nls_mod}{Object of class 'nls' representing the exponential model fit between FADs and time.  If this is not provided, the model will be estimated from the FADs provided or the FADs inferred from the provided tree.}

\item{showExpected}{whether or not to show the plot of the expected versus actual.  Default is FALSE}

\item{showTree}{Whether or not to plot the tree. Default is FALSE}
}
\description{
Bin FADs from a tree, and test for pulses using chi-square, following Vrba (1995)
}

