% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/DetectPulses.R
\name{detectPulses}
\alias{detectPulses}
\title{Test for pulses using using Foote Rates}
\usage{
detectPulses(treeDepth = 7, criterion = 1.5, nTaxa = 100, deathRate = 0,
  desiredBinNumber = 30, plotTree = FALSE, plotRates = TRUE)
}
\arguments{
\item{treeDepth}{The temporal depth (in millions of years) represented by the tree.}

\item{criterion}{How extreme an observed turnover rate must be to be considered a turnover pulse (relative to the interquartile distance for all bins considered as a group).}

\item{nTaxa}{The desired number of taxa in the simulated phylogenetic tree. Passed to phytools::pbtree.}

\item{deathRate}{The rate of extinction.  Passed to phytools::pbtree.}

\item{desiredBinNumber}{The number of bins that treeDepth will be broken into to compute turnover rates.}

\item{plotTree}{Whether or not to plot the simulated tree. Default is FALSE.}

\item{plotRates}{Whether or not to plot a histogram of the calculated rates.  Default is TRUE.}
}
\description{
Test for pulses using using Foote Rates
}

