% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectPulses.R
\name{detectPulses}
\alias{detectPulses}
\title{Test for pulses using using Foote Rates}
\usage{
detectPulses(birthRate, deathRate, totalTime = 7, criterion = 1.5,
  nTotalTaxa = c(100, 300), desiredBinNumber = 28, targetBins = seq(3.75,
  1.5, by = -0.25), plotTree = FALSE, plotRates = FALSE)
}
\arguments{
\item{birthRate}{The average (per interval, per lineage) rate of origination.}

\item{deathRate}{The average (per interval, per lineage) rate of extinction.}

\item{totalTime}{The temporal depth (in millions of years) represented by the simulation/}

\item{criterion}{How extreme an observed turnover rate must be to be considered a turnover pulse (relative to the interquartile distance for the distribution of all bins).}

\item{nTotalTaxa}{The minimum and maximum acceptable number of taxa in the tree.  A numeric vector of length two (e.g. c(min, max))}

\item{desiredBinNumber}{The number of bins for computing turnover rates. \}

\item{targetBins}{Vector indicating subset of bins in which to look for pulses. You must provide only the interval starting time.}

\item{plotTree}{Whether or not to plot the simulated tree. Default is FALSE.}

\item{plotRates}{Whether or not to plot a histogram of the calculated rates.  Default is TRUE.}
}
\description{
Test for pulses using using Foote Rates
}

