% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FitExponential.R
\name{FitExponential}
\alias{FitExponential}
\title{Internal Function to Fit exponential model to FADs}
\usage{
FitExponential(myTree, showPlot = FALSE)
}
\arguments{
\item{myTree}{an object of class "phylo"}

\item{showPlot}{whether or not to show plot.  Default is FALSE}

\item{showPlot}{Whether or not to show plots. Default is FALSE}
}
\value{
a Nonlinear Least Squares object of class "nls" representing the exponential fit
}
\description{
Internal Function to Fit exponential model to FADs
}

