% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRanges.R
\name{plotRanges}
\alias{plotRanges}
\title{plotRanges function plots fossil ranges using ggplot2
uses ggplot2 to plot ranges}
\usage{
plotRanges(fossilRange, returnValue = FALSE)
}
\description{
plotRanges function plots fossil ranges using ggplot2
uses ggplot2 to plot ranges
}

