% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPulses.R
\name{simRecords}
\alias{simRecords}
\title{test for pulses using paleotree functions}
\usage{
simRecords(origRateSim, extRateSim, nruns, totalTime, sampRateSim, startTaxa,
  maxAttempts, nSamp = c(0, 500), nExtant = c(10, 500), nTotalTaxa = c(10,
  500), plot = TRUE, print.runs = FALSE)
}
\description{
test for pulses using paleotree functions
}

